requirements.txt
# Основные библиотеки для YOLOv5
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.1
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.1.5
requests>=2.23.0
PyYAML>=5.3.1
tqdm>=4.64.0

# Для работы с изображениями
pillow>=8.0.0

# Для математических операций
numpy>=1.18.5
scipy>=1.4.1

# Для разметки данных (опционально)
# labelImg  # раскомментируйте если будете использовать LabelImg

# Для удобства
ipython
jupyter